refactor: separated star function & styles for category screen
