+++ src/components/ui/ProductComponent.js
@@ -1,8 +1,6 @@
-import React from "react";
 import { View, Text, Image, TouchableOpacity, StyleSheet } from "react-native";
-import Icon from "react-native-vector-icons/MaterialIcons";
 import { useNavigation } from "@react-navigation/native";
-import FractionalStar from "../ui//FractionalStar"; // Import your FractionalStar component
+import { renderStars } from "../../utils/renderStars/renderStars";
 
 const ProductComponent = ({ id, image, name, address, rating }) => {
   const navigation = useNavigation();
@@ -28,31 +26,10 @@ const ProductComponent = ({ id, image, name, address, rating }) => {
   );
 };
 
-const renderStars = (rating) => {
-  const stars = [];
-  const fullStars = Math.floor(rating);
-  const decimal = rating - fullStars;
 
-  for (let i = 0; i < fullStars; i++) {
-    stars.push(<Icon key={`full-${i}`} name="star" size={20} color="#FFD700" />);
-  }
-
-  if (decimal > 0) {
-    stars.push(<FractionalStar key="fractional" fill={decimal} />);
-  }
-
-  const totalStars = fullStars + (decimal > 0 ? 1 : 0);
-
-  for (let i = totalStars; i < 5; i++) {
-    stars.push(<Icon key={`empty-${i}`} name="star-border" size={20} color="#FFD700" />);
-  }
-
-  return stars;
-};
 
 const styles = StyleSheet.create({
   card: {
-    margin: 8,
     overflow: "hidden",
     borderRadius: 15,
   },
