+++ src/screens/CategoryScreen.js
@@ -1,7 +1,7 @@
 import { ActivityIndicator, FlatList, StyleSheet, Text } from "react-native";
 import ProductComponent from "../components/ui/ProductComponent";
 import { useRoute } from "@react-navigation/native";
-import { View } from "react-native";
+import { View, ScrollView } from "react-native";
 import { TopMenuComponent } from "../components/menu/TopMenuComponent";
 import { useSelector } from "react-redux";
 import { selectAuth } from "../store/slices/auth.slice";
@@ -37,12 +37,14 @@ const CategoryScreen = () => {
   const salonCount = filteredSalons.length;
 
   return (
-    <View contentContainerStyle={styles.container}>
+    <View>
       <TopMenuComponent title={categoryName} name={name} />
+      <ScrollView contentContainerStyle={styles.scrollView}>
       <HeadingComponent level={3} color="#000000" children={`${categoryName} - ${salonCount}`} />
       {filteredSalons.length > 0 ? (
         <FlatList
           data={filteredSalons}
+          ontentContainerStyle={styles.listContent}
           keyExtractor={(item) => item.salonId}
           renderItem={({ item }) => (
             <ProductComponent
@@ -50,21 +52,27 @@ const CategoryScreen = () => {
               image={item.image}
               name={item.title}
               address={item.description}
-              rating={item.rating || 0}
+              rating={item.rating || 3.5}
             />
           )}
         />
       ) : (
         <Text style={styles.emptyText}>No salons found for this category.</Text>
       )}
+      </ScrollView>
     </View>
   );
 };
 
 const styles = StyleSheet.create({
-  container: {
-    padding: 90,
+  scrollView:{
+    marginTop: 20,
+    marginHorizontal: 20,
+    gap: 20,
   },
+  listContent:{
+    gap: 20,
+  }
 });
 
 export default CategoryScreen;
